% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCSI.R
\name{calcCSI}
\alias{calcCSI}
\title{calcCSI}
\usage{
calcCSI(ter_data, GLNE_data, normalize = TRUE)
}
\arguments{
\item{ter_data}{A character vector with named elements representing terminal cell types and remaining cells (Other).}

\item{GLNE_data}{Gene local network energy matrix where columns represent cells and rows represent genes.}

\item{normalize}{A logical value indicating  whether the GLNE_data needs to be processed  (default is TRUE).}
}
\value{
A matrix representing the Cell Similarity Index (CSI) between remaining cells and terminal cell types.
Columns represent terminal cell types and rows represent remaining cells.
}
\description{
This function calculates the Cell Similarity Index (CSI) based on the input terminal data and gene local network energy (GLNE) data.
}
\examples{
# Example usage:
ter_data <- c("Cell1" = "Type1", "Cell2" = "Type2", "Cell3" = "Other") 
GLNE_data <- matrix(rnorm(100), nrow = 10, ncol = 10)  # Example GLNE data
result <- calcCSI(ter_data, GLNE_data, normalize =TRUE) # Names in ter_data must match the column names of GLNE_data.

}
