% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGEn.R
\name{calcGEn}
\alias{calcGEn}
\title{calcGEn}
\usage{
calcGEn(file_path, verbose = TRUE)
}
\arguments{
\item{file_path}{Path to the input CSV file.}

\item{verbose}{Logical. If TRUE, print progress messages. Default is TRUE.}
}
\value{
A list containing the gene local network energy matrix (GLNE) and a data frame with scEnergy and normalized scEnergy (scEn_Nor).
}
\description{
This function reads a CSV file containing single-cell gene expression data, performs Energy calculation using a Python script,
and returns the gene local network energy (GLNE) matrix and normalized scEnergy as a data frame.
}
\examples{
\dontrun{
# Example usage:
result <- calcGEn("path/to/your/file.csv")
}

}
