% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_terminal_states.R
\name{find_terminal_states}
\alias{find_terminal_states}
\title{find_terminal_states}
\usage{
find_terminal_states(
  Nor_en,
  theta = 0.85,
  decreasing = TRUE,
  diffvar = TRUE,
  maxPS = 3
)
}
\arguments{
\item{Nor_en}{A vector of normalized energy values for all cells. The name of the element is the index of the corresponding cell.}

\item{theta}{Threshold for ranking to identify terminal states. Default 0.85.}

\item{decreasing}{Logical value(TRUE or FALSE) indicating whether to rank in decreasing order, signifying a process in which cellular plasticity is reduced.}

\item{diffvar}{Logical value, parameter passed to PotencyFinder function.}

\item{maxPS}{Parameter passed to PotencyFinder function.}
}
\value{
A list containing cells with terminal states (`terminal_cells`) and filtered cells based on ranking (`candidate_cells`).
}
\description{
This function identifies terminal states in a numeric vector of normalized energy values for all cells.
The criteria for terminal state identification can be customized using various parameters.
}
\details{
The function calculates ranks based on the specified criteria and identifies terminal states accordingly.
It utilizes the PotencyFinder function to further analyze and filter potential states.
}
\examples{
# Example usage:
find_terminal_states(Nor_en_vector)

}
