% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPath.R
\name{findPath}
\alias{findPath}
\title{findPath}
\usage{
findPath(CSI, En_Nor, decreasing = TRUE)
}
\arguments{
\item{CSI}{A matrix containing cell state inference data. Output result of calcCSI function}

\item{En_Nor}{A data frame containing En_Nor data with columns 'scEn_Nor' and 'Class'.}

\item{decreasing}{A logical value indicating whether cellular plasticity should be decreasing (default is TRUE).}
}
\value{
A data frame On the basis of En_Nor, two columns of 'Path' and 'Pseudotime' are added, 
which correspond to the fate trajectory of the cell and the corresponding pseudotime value, respectively.
}
\description{
This function assigns fate paths to cells.
It calculates the category to which each cell belongs and creates a fate path.
}
\details{
This function performs fate assignment by finding the category to which each cell
belongs based on the maximum values in the CSI matrix. It then creates a fate path
and calculates pseudotime for each cell.
}
\examples{
# Example usage:
result <- findPath(CSI_data, En_Nor_data)

}
\seealso{
\code{\link{check_columns}}
}
