% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotencyFinder.R
\name{PotencyFinder}
\alias{PotencyFinder}
\title{PotencyFinder}
\usage{
PotencyFinder(
  Nor_en,
  pheno = NULL,
  diffvar = TRUE,
  maxPS = 5,
  verbose = FALSE,
  path = NULL
)
}
\arguments{
\item{Nor_en}{A vector of normalized energy values for all cells. The name of the element is the index of the corresponding cell.}

\item{pheno}{(Optional) A vector representing phenotypic information for each cell.}

\item{diffvar}{A logical value (TRUE or FALSE) indicating whether to assume different variances for clusters in the Gaussian Mixture Model.Default is TRUE.}

\item{maxPS}{An integer specifying the maximum number of potency states to infer.}

\item{verbose}{A logical value (TRUE or FALSE) indicating whether to display detailed output information.}

\item{path}{(Optional) A character string specifying the path to save output plots and model information.}
}
\value{
A list containing the Bayesian Information Criterion (BIC), probability matrix (proM), 
inferred potency states (potS), distribution of potency states across phenotypes (distPS), 
probability matrix for each phenotype (probPS), and normalised Heterogeneous Index for each phenotype (HI).
}
\description{
This function infers the discrete potency states of single cells and 
its distribution across the single cell population.
}
\examples{
# Example usage:
result <- PotencyFinder(Nor_en_vector, pheno_vector, diffvar = TRUE, maxPS = 5, verbose = TRUE, path = "output_path")

}
