% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotencyVisualizer.R
\name{PotencyVisualizer}
\alias{PotencyVisualizer}
\title{PotencyVisualizer}
\usage{
PotencyVisualizer(
  data,
  numGrid = 50,
  phi = 20,
  theta = 20,
  lphi = 30,
  k = 15,
  colPersp = NULL,
  colImage = NULL,
  keyPersp = list(length = 0.2, width = 0.4, shift = 0.15, cex.axis = 0.7, cex.clab =
    0.75),
  keyImage = list(length = 0.2, width = 0.4, shift = -0.15, cex.axis = 0.7, cex.clab =
    0.75),
  width = 9,
  height = 3,
  plotPath = NULL
)
}
\arguments{
\item{data}{A list containing elements 'coordinates' and 'potencyState'.}

\item{numGrid}{Number of grid points for density calculations.}

\item{phi}{Angle specifying the colatitude of the plot. See more details in ?plot3D::persp3D}

\item{theta}{Angle specifying theazimuthal direction of the plot.}

\item{lphi}{Lighting angle for perspective plots.}

\item{k}{Number of colors to be used in the color palette.}

\item{colPersp}{Color palette for perspective plots.Suggest the minimum number of colors to meet 2*Number of potential states}

\item{colImage}{Color palette for image plots. Suggest using continuous colors}

\item{keyPersp}{List of parameters for perspective plot color key.}

\item{keyImage}{List of parameters for image plot color key.}

\item{width}{Width of the output plot.}

\item{height}{Height of the output plot.}

\item{plotPath}{Path to save the output plot. if plotPath=NULL; Default is the current working directory}
}
\value{
None.
}
\description{
This function visualizes different potency states in 3D.
}
\examples{
# Example usage:
PotencyVisualizer(myData)

}
